test(ui): address notebook E2E review feedback
