+++ parish/apps/ui/e2e/fixtures.ts
@@ -7,7 +7,7 @@
  * helper dispatches to them.
  */
 
-import { test as base, type Page } from '@playwright/test';
+import { expect, test as base, type Page } from '@playwright/test';
 import {
 	SNAPSHOTS,
 	PALETTES,
@@ -37,6 +37,77 @@ export async function installTileRouteMock(page: Page): Promise<void> {
 	await page.route('**/tiles/**', (route) =>
 		route.fulfill({ status: 200, contentType: 'image/png', body: BLANK_PNG }),
 	);
+}
+
+/**
+ * Wait until Pixi has presented a texture-complete frame, not merely appended
+ * its canvas and emitted accessibility hit targets. The renderer loads assets
+ * asynchronously and the GPU upload/present can trail those DOM-ready signals;
+ * a raw Playwright screenshot taken in that gap contains large black texture
+ * rectangles. Sample the presented WebGL canvas from a requestAnimationFrame
+ * callback and fail closed unless the authored scene is both predominantly
+ * non-black and chromatically varied.
+ */
+export async function waitForTextureCompleteNotebookFrame(
+	page: Page,
+): Promise<void> {
+	const canvas = page
+		.getByTestId('illustrated-notebook-pixi-host')
+		.locator('canvas');
+
+	await expect
+		.poll(
+			() =>
+				canvas.evaluate(
+					(element) =>
+						new Promise<boolean>((resolve) => {
+							requestAnimationFrame(() => {
+								const source = element as HTMLCanvasElement;
+								const sample = document.createElement('canvas');
+								sample.width = 32;
+								sample.height = 20;
+								const context = sample.getContext('2d', {
+									willReadFrequently: true,
+								});
+								if (!context) {
+									resolve(false);
+									return;
+								}
+
+								context.drawImage(source, 0, 0, sample.width, sample.height);
+								const pixels = context.getImageData(
+									0,
+									0,
+									sample.width,
+									sample.height,
+								).data;
+								let nonBlack = 0;
+								const colourBuckets = new Set<number>();
+								for (let i = 0; i < pixels.length; i += 4) {
+									const red = pixels[i];
+									const green = pixels[i + 1];
+									const blue = pixels[i + 2];
+									const alpha = pixels[i + 3];
+									if (alpha > 0 && red + green + blue > 60) nonBlack += 1;
+									colourBuckets.add(
+										(red >> 4) * 256 + (green >> 4) * 16 + (blue >> 4),
+									);
+								}
+
+								const pixelCount = pixels.length / 4;
+								resolve(
+									nonBlack / pixelCount >= 0.8 && colourBuckets.size >= 20,
+								);
+							});
+						}),
+				),
+			{
+				message:
+					'Pixi notebook must present a texture-complete, non-degenerate frame',
+				timeout: 10_000,
+			},
+		)
+		.toBe(true);
 }
 
 /**
@@ -286,4 +357,4 @@ export const test = base.extend<{
 	},
 });
 
-export { expect } from '@playwright/test';
+export { expect };
